+++ CHANGES.rst
@@ -10,6 +10,10 @@ Unreleased
 -   Switched the minimum required version of :doc:`python-zyte-api
     <python-zyte-api:index>` from ``0.5.1`` to ``0.6.0``.
 
+-   Fixed the removal of default request headers (``Accept``,
+    ``Accept-Encoding``, ``Accept-Language``, and ``User-Agent``) not working
+    for request copies (e.g. redirects or retries).
+
 -   The default value of the :setting:`ZYTE_API_FALLBACK_HTTP_HANDLER` and
     :setting:`ZYTE_API_FALLBACK_HTTPS_HANDLER` settings is as expected even
     when not using the add-on.